ref(): move sources .cpp to src/ and .h to include/ for better code organization
