+++ src/aws.cpp
@@ -1,4 +1,4 @@
-#include "aws.h"
+#include "include/aws.h"
 #include <fstream>
 #include <string>
 #include <sstream>
