+++ tst_weather_ui.pro
@@ -9,11 +9,11 @@ CONFIG += c++14
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
-    aws.cpp \
     utils/CommandLineUtils.cpp \
-    data_base.cpp \
     driver/bmp280.cpp \
     driver/lis3mdl_reg.c \
+    src/data_base.cpp \
+    src/aws.cpp \
     src/RawData.cpp \
     src/bmp280data.cpp \
     src/data_manager.cpp \
@@ -24,17 +24,17 @@ SOURCES += \
     src/mqtt.cpp \
     src/raspberry_bmp280.cpp \
     src/settings.cpp\
-    settings_pop_up.cpp\
+    src/settings_pop_up.cpp\
     src/mki109v1.cpp \
 
 
 HEADERS += \
-    aws.h \
     utils/CommandLineUtils.h \
-    data_base.h \
     driver/bmp280.h \
     driver/bmp280_defs.h \
     driver/lis3mdl_reg.h \
+    include/aws.h \
+    include/data_base.h \
     include/RawData.h \
     include/bmp280data.h \
     include/data_manager.h \
@@ -44,9 +44,9 @@ HEADERS += \
     include/raspberry_bmp280.h \
     include/settings.h \
     include/bmp280data.h\
-    src/form.h\
-    settings_pop_up.h\
-    src/mki109v1.h \
+    include/form.h\
+    include/settings_pop_up.h\
+    include/mki109v1.h \
 
 FORMS += \
     src/form.ui \
